V3.10 - AGOL 5.1 Enhancements
Resolved following github tickets :

Enhancements:
1. #193 : Separate 'help' text available on splash screen from 'help'
available in-app
2. #211 : Configurable option for showing map first, instead of list
3. #221 : configurable option to load help as splash screen
4. #242 : Apply org theming
5. #252 : hide editing buttons when user doesn't have editing
permissions

Issues:
1. #234 : RtL issues
2. #235 : missing outline on 'my location' button
