+++ config/defaults.js
@@ -97,6 +97,8 @@ define({
     "enableFeatureDelete" : false,
     "reportedByField": "USERID",
     "locationField": "LONG_TEXT",
+    "showMapFirst":false,
+    "showHelpOnLoad": false,
 
     "zoomLevel": 12,
     "enableUSNGSearch": false,
@@ -114,6 +116,9 @@ define({
     "helpLinkText": "Help",
     "helpDialogTitle":"Help",
     "helpDialogContent": "<p>Crowdsource Reporter is a configurable gallery application template that allows users to submit problems or observations. The application has been optimized for smartphones but is responsively designed to be used on smartphones, tablets, and desktop computers.</p><p>The Crowdsource Reporter application presents one or more maps that can be used to report a problem or observation. Users can anonymously submit new reports, review existing reports, and comment and vote on reports or observations submitted by other users. They also can authenticate with their social media or ArcGIS Online credentials and track the status of problems or observations they have reported.</p>",
+    "enableDifferentHelpContent":false,
+    "loginHelpDialogTitle":"",
+    "loginHelpDialogContent": "",
     "bufferRadius": 3, //Buffer radius will be used to create inital buffer on applicaiton load
     "bufferUnit": "miles",
     "geoformDetailsSectionLabel": "Details", //label for details section in geoform
