+++ index.html
@@ -50,7 +50,7 @@
                     <div class="esriCTHeaderText esriCTMapHeaderText esriCTEllipsis" id="mapContainerTitle"></div>
                     <div class="esriCTListViewButton esriCTPointerCursor icon icon-list" id="toggleListViewButton"></div>
                 </div>
-                <div id="mapContainer">
+                <div id="mapContainer" class="esriCTMap">
                     <!-- The ArcGIS API for JavaScript provides bidirectional support.  When viewing the application in an right to left (rtl) language like Hebrew and Arabic the map needs to remain in left-to-right (ltr) mode. Specify this by setting the dir attribute on the div to ltr. -->
                     <div id="mapDiv" dir="ltr" data-dojo-props="region:'center'">
                     </div>
@@ -68,7 +68,7 @@
             </div>
         </div>
 
-        <div id="geoformContainer" class="esriCTHidden esriCTBorderLeft esriCTGeoFormParentContainer">
+        <div id="geoformContainer" class="esriCTHidden esriCTBorderLeft esriCTGeoFormParentContainer esriCTHeaderTextColorAsBorder">
         </div>
         <div id="mobileMenuContainer" class="esriCTMobileMenuContainer esriCTHideMobileMenu">
         </div>
