+++ js/main.js
@@ -32,7 +32,6 @@ define([
     "dojo/window",
     "dojo/aspect",
     "dojo/Deferred",
-    "dojo/text!css/theme-template.css",
     "esri/layers/GraphicsLayer",
     "esri/layers/FeatureLayer",
     "esri/geometry/Circle",
@@ -80,7 +79,6 @@ define([
     dojowindow,
     aspect,
     Deferred,
-    ThemeCss,
     GraphicsLayer,
     FeatureLayer,
     Circle,
@@ -152,7 +150,7 @@ define([
             // config will contain application and user defined info for the template such as i18n strings, the web map id
             // and application id
             // any url parameters and any application specific configuration information.
-            var queryParams = {}, rgbColor;
+            var queryParams = {};
             if (boilerPlateTemplateObject) {
                 this.boilerPlateTemplate = boilerPlateTemplateObject;
                 this.config = boilerPlateTemplateObject.config;
@@ -163,44 +161,6 @@ define([
                 this.appUtils = new ApplicationUtils({
                     "config": this.config
                 });
-                if (!this.config.portalProperties) {
-                    rgbColor = new Color(this.config.theme);
-                    rgbColor.a = 0.6;
-                    this.config.portalProperties = {
-                        "links": {},
-                        "showSocialMediaLinks": true,
-                        "sharedTheme": {
-                            "header": {
-                                "background": this.config.theme,
-                                "text": "#fff"
-                            },
-                            "body": {
-                                "background": "#FFF",
-                                "text": "#333"
-                            },
-                            "button": {
-                                "background": "#D95F45",
-                                "text": "#333"
-                            },
-                            "logo": {
-                                "small": ""
-                            },
-                            calculatedBodyBackgroundColor: "#FFF",
-                            calculatedBodyTextColor: "#f0f0f0",
-                            calculatedHeaderBackgroundColor: rgbColor,
-                            calculatedHeaderTextColor: "gray"
-                        }
-                    };
-                } else {
-                    this.config.portalProperties.sharedTheme.calculatedBodyBackgroundColor =
-                        this.appUtils.getCalculatedColor(this.config.portalProperties.sharedTheme.body.background, 50, 6);
-                    this.config.portalProperties.sharedTheme.calculatedBodyTextColor =
-                        this.appUtils.getCalculatedColor(this.config.portalProperties.sharedTheme.body.text, 50, 21);
-                    this.config.portalProperties.sharedTheme.calculatedHeaderBackgroundColor =
-                        this.appUtils.getCalculatedColor(this.config.portalProperties.sharedTheme.header.background, 70, 18);
-                    this.config.portalProperties.sharedTheme.calculatedHeaderTextColor =
-                        this.appUtils.getCalculatedColor(this.config.portalProperties.sharedTheme.header.text, 50, 27);
-                }
 
                 // Initializes the map-search widget
                 this.mapSearch = new MapSearch({ "config": this.config, "appUtils": this.appUtils, "handleFeatureSearch": true });
@@ -244,6 +204,7 @@ define([
                     };
                     this._menusList.signOut = true;
                     this._menusList.signIn = false;
+                    this.config.logInDetails.canEditFeatures = loggedInUser.canEditFeatures;
                 } else {
                     this.config.logInDetails = {
                         "userName": "",
@@ -252,6 +213,8 @@ define([
                     };
                     this._menusList.signIn = true;
                     this._menusList.signOut = false;
+                    //If user is not logged in keep editing flag to true by default
+                    this.config.logInDetails.canEditFeatures = true;
                 }
                 //By default we have disabled queryForGroupItems
                 //since it was getting group items for the group configured in default.js only,
@@ -327,10 +290,6 @@ define([
             if (!this.config.showNullValueAs) {
                 this.config.showNullValueAs = "";
             }
-
-            //Set Application Theme
-            this._loadApplicationTheme();
-
             //Set Application header
             this._createApplicationHeader();
 
@@ -445,7 +404,6 @@ define([
                         this._updateFeatureInIssueWall(this._itemDetails.item, false, true);
                     }
                 });
-
                 var submitButtonText, submitButtonColor;
                 if (this.config && lang.trim(this.config.submitReportButtonText) === "") {
                     submitButtonText = this.config.i18n.main.submitReportButtonText;
@@ -457,7 +415,11 @@ define([
                 domStyle.set(dom.byId("submitFromMap"), "background-color", submitButtonColor);
 
                 on(dom.byId("submitFromMap"), "click", lang.hitch(this, function (evt) {
-                    this._createGeoForm();
+                    if (this.config.logInDetails.canEditFeatures) {
+                        this._createGeoForm();
+                    } else {
+                        this.appUtils.showMessage(this.config.i18n.main.noEditingPermissionsMessage);
+                    }
                 }));
                 on(dom.byId("mapBackButton"), "click", lang.hitch(this, function (evt) {
                     this._toggleListView();
@@ -636,193 +598,6 @@ define([
                 domAttr.set(dom.byId("noWebMapChildDiv"), "innerHTML", noMapMessage);
             } catch (err) {
                 this.appUtils.showError(err.message);
-            }
-        },
-
-        /**
-        * This function rturns the app theme object for backward compatibility
-        * @memberOf main
-        */
-        _getDefaultAppTheme: function () {
-            var defaultTheme;
-            defaultTheme = {
-                "header": {
-                    "background": "#fff",
-                    "text": this.config.theme,
-                    "calculatedBackground": "#3daceb",
-                    "calculatedText": this.appUtils.getCalculatedColor(this.config.theme, 70, 18)
-                },
-                "body": {
-                    "background": "#fff",
-                    "text": "#515151",
-                    "calculatedBackground": "#f0f0f0",
-                    "calculatedText": "#878787"
-                },
-                "button": {
-                    "background": "#fff",
-                    "text": this.config.theme
-                }
-            };
-            return defaultTheme;
-        },
-
-        /**
-        * Set's the default theme properties for empty colors in shared theme
-        * @memberOf main
-        */
-        _updateForEmptyColors: function (sharedTheme, defaultTheme) {
-            if (!sharedTheme) {
-                sharedTheme = {};
-            }
-            if (!sharedTheme.header) {
-                sharedTheme.header = {};
-            }
-            if (!sharedTheme.body) {
-                sharedTheme.body = {};
-            }
-            if (!sharedTheme.button) {
-                sharedTheme.button = {};
-            }
-            if (!sharedTheme.logo) {
-                sharedTheme.logo = {};
-            }
-            if (!sharedTheme.header.background) {
-                sharedTheme.header.background = defaultTheme.header.background;
-            }
-            if (!sharedTheme.header.text) {
-                sharedTheme.header.text = defaultTheme.header.text;
-            }
-            if (!sharedTheme.body.background) {
-                sharedTheme.body.background = defaultTheme.body.background;
-            }
-            if (!sharedTheme.body.text) {
-                sharedTheme.body.text = defaultTheme.body.text;
-            }
-           if (!sharedTheme.button.background) {
-                sharedTheme.button.background = defaultTheme.button.background;
-            }
-            if (!sharedTheme.button.text) {
-                sharedTheme.button.text = defaultTheme.button.text;
-            }
-            if (!sharedTheme.logo.small && defaultTheme.logo && defaultTheme.logo.small) {
-                sharedTheme.logo.small = defaultTheme.logo.small;
-            }
-            return sharedTheme;
-        },
-
-        /**
-        * Set's the org theme properties for the app theme
-        * This function will create dafault app theme if appTheme is not available in config
-        * @memberOf main
-        */
-        _setOrgTheme: function () {
-            var sharedTheme, calculateColors, defaultThemeSettings;
-            calculateColors = false;
-            //get default theme settings using only previous single color configuration
-            defaultThemeSettings = this._getDefaultAppTheme();
-            if (this.config.portalObject && this.config.portalObject.portalProperties &&
-                this.config.portalObject.portalProperties.sharedTheme) {
-                sharedTheme = this.config.portalObject.portalProperties.sharedTheme;
-                //Update EMPTY values of shard theme with default values if any
-                sharedTheme = this._updateForEmptyColors(sharedTheme, defaultThemeSettings);
-            }
-            //Check if user has configured any values using configuration panel
-            if (this.config.headerBackgroundColor) {
-                this.config.appTheme = {
-                    "header": {
-                        "background": this.config.headerBackgroundColor,
-                        "text": this.config.headerTextColor
-                    },
-                    "body": {
-                        "background": this.config.bodyBackgroundColor,
-                        "text": this.config.bodyTextColor
-                    },
-                    "button": {
-                        "background": this.config.buttonBackgroundColor,
-                        "text": this.config.buttonTextColor
-                    }
-                }
-                //Update empty values in app theme
-                if (sharedTheme) {
-                    this.config.appTheme = this._updateForEmptyColors(this.config.appTheme, sharedTheme);
-                }
-                calculateColors = true;
-            } else if (sharedTheme) {
-                //Mixin shared theme in appTheme
-                this.config.appTheme = sharedTheme;
-                //Calculate colors as we are getting values from shared theme
-                calculateColors = true;
-            } else {
-                this.config.appTheme = defaultThemeSettings;
-            }
-            //if logo is not configured by user and in org properties we have valid logo then only use the logo from org
-            if (!this.config.applicationIcon && this.config.appTheme.logo && this.config.appTheme.logo.small) {
-                this.config.applicationIcon = this.config.appTheme.logo.small;
-            }
-            //if don't have calculated colors calculate it
-            if (this.config.appTheme && calculateColors) {
-                this.config.appTheme.body.calculatedBackground =
-                    this.appUtils.getCalculatedColor(this.config.appTheme.body.background, 50, 6);
-                this.config.appTheme.body.calculatedText =
-                    this.appUtils.getCalculatedColor(this.config.appTheme.body.text, 50, 21);
-                this.config.appTheme.header.calculatedBackground =
-                    this.appUtils.getCalculatedColor(this.config.appTheme.header.background, 70, 18);
-                this.config.appTheme.header.calculatedText =
-                    this.appUtils.getCalculatedColor(this.config.appTheme.header.text, 50, 27);
-            }
-        },
-
-        /**
-        * Load application theme
-        * @memberOf main
-        */
-        _loadApplicationTheme: function () {
-            var cssString, head, style, link;
-            //if theme is configured
-            if (this.config.theme) {
-                //set the orgtheme for application
-                this._setOrgTheme();
-                //substitute color values in theme template
-                cssString = string.substitute(ThemeCss, {
-                    SelectedThemeColor: this.config.theme,
-                    // Configured/Org colors for app theme
-                    BodyBackgroundColor: this.config.appTheme.body.background,
-                    BodyTextColor: this.config.appTheme.body.text,
-                    HeaderBackgroundColor: this.config.appTheme.header.background,
-                    HeaderTextColor: this.config.appTheme.header.text,
-                    ButtonBackgroundColor: this.config.appTheme.button.background,
-                    ButtonTextColor: this.config.appTheme.button.text,
-                    // Calculated colors
-                    CalculatedBodyBackgroundColor: this.config.appTheme.body.calculatedBackground,
-                    CalculatedBodyTextColor: this.config.appTheme.body.calculatedText,
-                    CalculatedHeaderBackgroundColor: this.config.appTheme.header.calculatedBackground,
-                    CalculatedHeaderTextColor: this.config.appTheme.header.calculatedText,
-                    OverlayBackgroundColor:
-                        this.appUtils.getOverlayBackgroundColor(this.config.appTheme.header.text)
-                });
-                //Create Style using theme template and append it to head
-                //On Lower versions of IE10 Style tag is read only so create theme using styleSheet.cssText
-                if (dojo.isIE < 10) {
-                    head = document.getElementsByTagName('head')[0];
-                    style = document.createElement('style');
-                    style.type = 'text/css';
-                    style.styleSheet.cssText = cssString;
-                    head.appendChild(style);
-                } else {
-                    domConstruct.create("style", {
-                        "type": "text/css",
-                        "innerHTML": cssString
-                    }, query("head")[0]);
-                }
-
-                //If application is loaded in RTL mode, change styles of required nodes
-                if (this.config.i18n.direction === "rtl") {
-                    link = document.createElement('link');
-                    link.rel = 'stylesheet';
-                    link.type = 'text/css';
-                    link.href = "./css/rtl.css";
-                    document.getElementsByTagName('head')[0].appendChild(link);
-                }
             }
         },
 
@@ -1299,6 +1074,10 @@ define([
                     this._sidebarCnt.showPanel("issueWall");
                 }
             } else {
+                //Check for the configurable parameter and accordingly show map first in mobile devices
+                if (this.config.showMapFirst && dojowindow.getBox().w < 768) {
+                    this._toggleMapView();
+                }
                 data.featureLayerCount = this.featureLayerCount;
                 data.layerGraphicsArray = this.layerGraphicsArray;
                 this._issueWallWidget.initIssueWall(data);
